Extract convertWorkflowToUseSteps Move convertWorkflowToUseSteps into a helper function
