+++ packages/lib-classifier/src/store/WorkflowStepStore/helpers/index.js
@@ -1,0 +1,1 @@
+export { default as convertWorkflowToUseSteps } from './convertWorkflowToUseSteps'
